--- internal/util-cache/src/main/scala/sbt/internal/util/CacheImplicits.scala
@@ -1,7 +1,0 @@
-package sbt.internal.util
-
-import sjsonnew.BasicJsonProtocol
-
-object CacheImplicits extends BasicCacheImplicits
-  with BasicJsonProtocol
-  with HListFormat
