+++ util-cache/src/main/scala/sbt/util/CacheImplicits.scala
@@ -1,0 +1,9 @@
+package sbt.util
+
+import sjsonnew.BasicJsonProtocol
+import sbt.internal.util.HListFormats
+
+object CacheImplicits extends CacheImplicits
+trait CacheImplicits extends BasicCacheImplicits
+  with BasicJsonProtocol
+  with HListFormats
